refactor: move analytic utils from lo to course
